refactor: move Make.racingInfo to ConvertInputTo
Sub Make.racingInfo to ConvertInputTo
Delete Make because there is no function

There's no reason to put a function that creates a racingInfo in a Make module
