+++ src/App.js
@@ -1,8 +1,8 @@
-import { Get, Make, Print } from './modules/AppModules.js';
+import { ConvertInputTo, Get, Print } from './modules/AppModules.js';
 
 class App {
   async play() {
-    const racingInfo = await Make.racingInfo();
+    const racingInfo = await ConvertInputTo.racingInfo();
     Print.racingResultFrom(racingInfo);
     Print.winnerFrom(Get.winnerNameList(racingInfo));
   }
