+++ src/modules/AppModules.js
@@ -2,16 +2,14 @@ import ConvertInputTo from './ConvertInputTo.js';
 import ErrorCheck from './ErrorCheck.js';
 import Get from './Get.js';
 import Is from './Is.js';
-import Make from './Make.js';
 import Print from './Print.js';
 
-export { ConvertInputTo, ErrorCheck, Get, Is, Make, Print };
+export { ConvertInputTo, ErrorCheck, Get, Is, Print };
 
 export const AppModules = {
   ConvertInputTo,
   ErrorCheck,
   Get,
   Is,
-  Make,
   Print,
 };
